Moves polling back to device
